ai-processor/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.8.3
pyyaml>=6.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
